+++ src/LinksComponents/SearchBar.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import searchIcon from '@/assets/icons/ic_search.svg';
 import closeIcon from '@/assets/images/close.png';
 
-export const SearchBar = () => {
+const SearchBar = () => {
   return (
     <div className="m-auto max-w-[1060px] h-auto px-4 py-[15px] flex justify-between items-center rounded-[10px] bg-[#f5f5f5]">
       <div className="w-full flex justify-center items-center gap-3">
@@ -19,3 +19,5 @@ export const SearchBar = () => {
     </div>
   );
 };
+
+export default SearchBar;
